git: fix communication protocol.
git-remote-gittorrent:
Only fetch a reference when git sends a "fetch" command. It knows when it
already has a commit and won't ask for it. If it asks for nothing, quit early.

Store references in a dictionary so that when asked for a specific sha, we can
find its reference quickly.

gittorrentd:
Add branches to `userProfile`, even if that sha has already been announced.

Fix the parsing of `git-upload-pack`'s output to not skip the line pointing to
`HEAD`.